requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.2
cssselect==1.0.3
Flask==0.12.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.4
Flask-GraphQL==1.4.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.6.7
graphene==1.4.1
graphene-sqlalchemy==1.1.1
graphql-core==1.1
graphql-relay==0.4.5
gspread==3.0.0
httplib2==0.11.3
idna==2.6
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.9.6
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
oauth2client==4.1.2
pluggy==0.3.1
promise==2.0.2
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyMySQL==0.8.0
pyquery==1.4.0
python-dateutil==2.7.3
python-dotenv==0.7.1
python-editor==1.0.3
PyYAML==3.11
readability-lxml==0.6.2
requests==2.18.4
rsa==3.4.2
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.11
SQLAlchemy-Paginator==0.1
tox==2.2.1
typing==3.6.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.12.2
